Little bug with button disabled coloring
The color intended to display when the button is displayed shows up even when the correct answer is chosen, which is technically correct as we have set the button to disable even on correct answers to prevent duplicate points for one question. The problem is how to preven this coloring on correct answers though.